requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
bitarray==2.9.2
bitstring==4.2.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.10.70
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
defusedxml==0.7.1
ecdsa==0.19.0
esp-coredump==1.12.0
esp-idf-kconfig==2.3.0
esp-idf-monitor==1.5.0
esp-idf-panic-decoder==1.2.1
esp-idf-size==1.6.0
esp_idf_nvs_partition_gen==0.1.3
esptool==4.8.1
flatbuffers==24.3.25
fonttools==4.53.1
fpdf2==2.8.1
freertos-gdb==1.0.3
gast==0.6.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.68.1
guidata==3.1.0
guiqwt==4.4.4
h5py==3.12.1
idf-component-manager==2.0.4
idna==3.10
imutils==0.5.4
intelhex==2.3.0
jax==0.4.31
jaxlib==0.4.31
jsonref==1.1.0
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.2.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.13.1
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclang==0.5.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyelftools==0.31
pygame==2.6.0
pygdbmi==0.11.0.0
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
pyqt-tools==1.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PythonQwt==0.14.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
QtPy==2.4.2
reedsolo==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rsa==4.9
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tk==0.1.0
tomli==2.2.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
wrapt==1.14.1